Fix invalid memory read in system tests
The test test_CreateObject_RSA defines a buffer (named modulus) of size
257 on the stack: CK_BYTE modulus[ 257 ] = { 0 };
The test wants to pass the address of second byte in the buffer
to the C_CreateObject function. The test code was incorrectly using the
expression &modulus + 1 which was resulting in address of the memory
right after the buffer. This commits updates the expression to
&modulus[ 1 ] to correctly use the address of the second byte.

This commit also fixes a memory leak by ensuring that the memory
allocated in the C_VerifyInit is freed in C_Verify.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>